[improve] code organizing
