+++ src/pages/Control/Control.tsx
@@ -1,12 +1,13 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import Loading from '../../components/Loading';
-import { Base } from '../../utils/base';
-import { Matrix } from '../../utils/common';
-import { createSearch } from '../../utils/layout';
-import { renderMatrix } from '../../utils/layout/render';
-import { closeAllWindow, SearchWindow } from '../../utils/layout/window';
-import { getSearchword } from '../../utils/search';
-import ArrowButtonGroup from './components/ArrowGroup';
+import React, { useCallback, useEffect, useState } from 'react'
+import Loading from '../../components/Loading'
+import { Base } from '../../utils/base'
+import { Matrix } from '../../utils/common'
+import { calcControlWindowPos } from '../../utils/control-window'
+import { createSearch } from '../../utils/layout'
+import { renderMatrix } from '../../utils/layout/render'
+import { closeAllWindow, SearchWindow } from '../../utils/layout/window'
+import { getSearchword } from '../../utils/search'
+import ArrowButtonGroup from './components/ArrowGroup'
 
 import SearchForm from './components/SearchForm'
 
@@ -81,7 +82,7 @@ const ControlApp: React.FC<{ base: Base }> = ({ base }) => {
   const moveControlWindow = useCallback(async () => {
     const { id } = await chrome.windows.getCurrent()
     if (id !== undefined) {
-      const [top, left] = base.calcControlWindowPos()
+      const [ top, left ] = calcControlWindowPos(base.total_height, base.limit)
       await chrome.windows.update(id, { top, left })
     }
   }, [base])
